test: Move assertions+builders+doubles+helpers to support/ subdir
