+++ test/integration/case_or_encoding_change.js
@@ -10,10 +10,10 @@ import {
 } from 'mocha'
 import should from 'should'
 
-import configHelpers from '../helpers/config'
-import * as cozyHelpers from '../helpers/cozy'
-import pouchHelpers from '../helpers/pouch'
-import { IntegrationTestHelpers } from '../helpers/integration'
+import configHelpers from '../support/helpers/config'
+import * as cozyHelpers from '../support/helpers/cozy'
+import pouchHelpers from '../support/helpers/pouch'
+import { IntegrationTestHelpers } from '../support/helpers/integration'
 
 suite('Case or encoding change', () => {
   if (process.env.TRAVIS && (process.platform === 'darwin')) {
