+++ test/integration/conflict_resolution.js
@@ -11,10 +11,10 @@ import {
 import should from 'should'
 import sinon from 'sinon'
 
-import configHelpers from '../helpers/config'
-import * as cozyHelpers from '../helpers/cozy'
-import pouchHelpers from '../helpers/pouch'
-import { IntegrationTestHelpers } from '../helpers/integration'
+import configHelpers from '../support/helpers/config'
+import * as cozyHelpers from '../support/helpers/cozy'
+import pouchHelpers from '../support/helpers/pouch'
+import { IntegrationTestHelpers } from '../support/helpers/integration'
 
 suite('Conflict resolution', () => {
   let cozy, helpers
