+++ test/integration/scenarios.js
@@ -8,11 +8,11 @@ import path from 'path'
 import should from 'should'
 import sinon from 'sinon'
 
-import { scenarios, loadFSEventFiles, runActions, init } from '../helpers/scenarios'
-import configHelpers from '../helpers/config'
-import * as cozyHelpers from '../helpers/cozy'
-import { IntegrationTestHelpers } from '../helpers/integration'
-import pouchHelpers from '../helpers/pouch'
+import { scenarios, loadFSEventFiles, runActions, init } from '../support/helpers/scenarios'
+import configHelpers from '../support/helpers/config'
+import * as cozyHelpers from '../support/helpers/cozy'
+import { IntegrationTestHelpers } from '../support/helpers/integration'
+import pouchHelpers from '../support/helpers/pouch'
 import remoteCaptureHelpers from '../../dev/capture/remote'
 
 let helpers
