+++ test/support/assertions/timestamp.js
@@ -1,6 +1,6 @@
 import should from 'should'
 
-import timestamp from '../../core/timestamp'
+import timestamp from '../../../core/timestamp'
 
 should.use(function (should, Assertion) {
   Assertion.add('sameTimestamp', function (expected, message) {
