+++ test/support/builders/metadata/dir.js
@@ -1,22 +1,26 @@
-/* @flow */
+// @flow
 
+import { assignId } from '../../../../core/metadata'
 import BaseMetadataBuilder from './base'
-import { assignId } from '../../../core/metadata'
 
-import type { Metadata } from '../../../core/metadata'
+import type { Metadata } from '../../../../core/metadata'
 
 import pouchdbBuilders from '../pouchdb'
 
-export default class FileMetadataBuilder extends BaseMetadataBuilder {
+export default class DirMetadataBuilder extends BaseMetadataBuilder {
   build (): Metadata {
     const doc = {
       ...this.opts,
       _id: '',
       // _rev: pouchdbBuilders.rev(),
-      docType: 'file',
+      docType: 'folder',
       remote: {
         _id: pouchdbBuilders.id(),
         _rev: pouchdbBuilders.rev()
+      },
+      sides: {
+        local: 1,
+        remote: 1
       },
       tags: [],
       updated_at: '2017-06-08T15:09:52.000Z'
