+++ test/support/builders/stream.js
