+++ test/support/doubles/cozy_stack.js
