+++ test/support/helpers/integration.js
@@ -5,14 +5,14 @@ import EventEmitter from 'events'
 import { pick } from 'lodash'
 import sinon from 'sinon'
 
-import Config from '../../core/config'
-import Ignore from '../../core/ignore'
-import Local from '../../core/local'
-import Merge from '../../core/merge'
-import Pouch from '../../core/pouch'
-import Prep from '../../core/prep'
-import Remote from '../../core/remote'
-import Sync from '../../core/sync'
+import Config from '../../../core/config'
+import Ignore from '../../../core/ignore'
+import Local from '../../../core/local'
+import Merge from '../../../core/merge'
+import Pouch from '../../../core/pouch'
+import Prep from '../../../core/prep'
+import Remote from '../../../core/remote'
+import Sync from '../../../core/sync'
 
 import { LocalTestHelpers } from './local'
 import { RemoteTestHelpers } from './remote'
