+++ test/unit/remote/watcher.js
@@ -9,11 +9,11 @@ import sinon from 'sinon'
 import should from 'should'
 import { Client as CozyClient } from 'cozy-client-js'
 
-import pouchdbBuilders from '../../builders/pouchdb'
-import configHelpers from '../../helpers/config'
-import { onPlatform } from '../../helpers/platform'
-import pouchHelpers from '../../helpers/pouch'
-import { builders } from '../../helpers/cozy'
+import pouchdbBuilders from '../../support/builders/pouchdb'
+import configHelpers from '../../support/helpers/config'
+import { onPlatform } from '../../support/helpers/platform'
+import pouchHelpers from '../../support/helpers/pouch'
+import { builders } from '../../support/helpers/cozy'
 
 import { createMetadata } from '../../../core/conversion'
 import * as metadata from '../../../core/metadata'
